requirements.txt
streamlit==1.28.1
requests==2.31.0
psycopg2-binary==2.9.7
sqlalchemy==2.0.21
pandas==2.1.1
python-dotenv==1.0.0
Pillow==10.0.1
